% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SoilTesting.R
\docType{data}
\name{df_Available_K}
\alias{df_Available_K}
\title{Example Data Frame for available potassium (K) in soils}
\format{
Data frame with Mass of soil sample (g) in the first column, Volume of extractant solution (mL) in the second column, Volume of aliquot of the extract (mL) in the third column, Volume of final solution fed to flame photometer (mL) in the fourth column, K concentration (mg/L) in final solution as obtained directly from the flame photometer or indirectly from the ‘Std_Curve’ function in the fifth column. Write the following notations on the spreadsheet:

Mass_Soil - for Mass of soil sample (g)

Vol_Extractant - for Volume of extractant solution (mL)

Vol_Aliquot - for Volume of aliquot of the extract (mL)

Vol_Final - for Volume of final solution fed to flame photometer (mL)

Flame_Reading - for K concentration (mg/L) in final solution as obtained directly from the flame photometer or indirectly from the ‘Std_Curve’ function
In case of no additional dilution after extraction, the values of Vol_Aliquot and Vol_Final should be put as 1.
}
\usage{
df_Available_K
}
\description{
User is advised to prepare the data as suggested in the example to get the content of available K in soil based on flame photometer reading.
}
\keyword{datasets}
