% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SoilTesting.R
\docType{data}
\name{df_Available_P}
\alias{df_Available_P}
\title{Example Data Frame for available phosphorus (P) in soils}
\format{
Data frame with Mass of soil sample (g) in the first column, Volume of extractant solution (mL) in the second column, Volume of aliquot of the extract (mL) in the third column, Volume of coloured solution (mL) in the fourth column, P concentration (mg/L) in coloured solution as obtained from the ‘Std_Curve’ function in the fifth column. Write the following notations on the spreadsheet:

Mass_Soil - for Mass of soil sample (g)

Vol_Extractant - for Volume of extractant solution (mL)

Vol_Aliquot - for Volume of aliquot of the extract (mL)

Vol_Colour - for Volume of coloured solution (mL)

P_mgL_Std_curve - for P concentration in coloured solution (mg/L) as obtained from the ‘Std_Curve’ function
}
\usage{
df_Available_P
}
\description{
User is advised to prepare the data as suggested in the example to get the content of available P in soil based on spectrophotometric determination using ascorbic acid blue colour method (Watanabe and Olsen, 1965).
}
\keyword{datasets}
