% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SoilTesting.R
\docType{data}
\name{df_DTPA_Micro}
\alias{df_DTPA_Micro}
\title{Example Data Frame for available micronutrient cations in soils}
\format{
Data frame with Mass of soil sample (g) in the first column, Volume of extractant solution (mL) in the second column, Volume of aliquot of the extract (mL) in the third column, Volume of final solution fed to AAS (mL) in the fourth column, Concentration of the micronutrient cation (mg/L) in final solution (test sample) as obtained from the AAS in the fifth column, Concentration of the micronutrient cation (mg/L) in the blank solution as obtained from the AAS in the sixth column. Write the following notations on the spreadsheet:

Mass_Soil - for Mass of soil sample (g)

Vol_Extractant - for Volume of extractant solution (mL)

Vol_Aliquot - for Volume of aliquot of the extract (mL)

Vol_Final - for Volume of final solution fed to AAS (mL)

AAS_Sample	for Concentration of the micronutrient cation (mg/L) in final solution (test sample) as obtained from the AAS

AAS_Blank	for Concentration of the micronutrient cation (mg/L) in the blank solution as obtained from the AAS

In case of no additional dilution after extraction, the values of Vol_Aliquot and Vol_Final should be put as 1.
}
\usage{
df_DTPA_Micro
}
\description{
User is advised to prepare the data as suggested in the example to get the content of available micronutrient cations in soil based on atomic absorption spectrophotometer (AAS) reading.
}
\keyword{datasets}
