% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SoilTesting.R
\docType{data}
\name{df_Minearl_N}
\alias{df_Minearl_N}
\title{Example Data Frame for mineral nitrogen (N) in soil}
\format{
Data frame with Mass of moist soil sample (g) in the first column, Moisture content (\% w/w) in the second column, Volume of aliquot of the extract (mL) taken for distillation in the third column, Volume of 2 M KCl (mL) added as extractant in the fourth column, Volume of standard sulphuric acid solution (mL) consumed in titration with soil in the fifth column, Volume of standard sulphuric acid solution (mL) consumed in titration without soil (i.e., blank titration) in the sixth column, and Normality of standard sulphuric acid solution in the seventh column. Write the following notations on the spreadsheet:

Mass_Moistsoil - for Mass of moist soil sample (g)

Moisture_Percent - for Moisture content of the field-moist soil sample (\% w/w)

Vol_Aliquot - for Volume of aliquot of the extract (mL) taken for distillation

Vol_Extractant - for Volume of 2 M KCl (mL) added as extractant

Vol_Sulphuric_Soil - for Volume of standard sulphuric acid solution (mL) consumed in titration with soil

Vol_Sulphuric_Blank - for Volume of standard sulphuric acid solution (mL) consumed in titration without soil (i.e., blank titration)

Normality_Sulphuric - for Normality of standard sulphuric acid solution
}
\usage{
df_Minearl_N
}
\description{
User is advised to prepare the data as suggested in the example to get the mineral N (ammonium-N and nitrate-N together) in soil
}
\keyword{datasets}
