% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SoilTesting.R
\docType{data}
\name{df_N_Boric}
\alias{df_N_Boric}
\title{Example Data Frame for total nitrogen (N) and available or mineralizable N in soil}
\format{
Data frame with Mass of soil sample (g) in the first column, Volume of standard sulphuric acid solution (mL) consumed in titration with soil in the second column, Volume of standard sulphuric acid solution (mL) consumed in titration without soil (i.e., blank titration) in the third column, and Normality of standard sulphuric acid solution in the fourth column. Write the following notations on the spreadsheet:

Mass_Soil - for Mass of soil sample (g)

Vol_Sulphuric_Soil - for Volume of standard sulphuric acid solution (mL) consumed in titration with soil

Vol_Sulphuric_Blank - for Volume of standard sulphuric acid solution (mL) consumed in titration without soil (i.e., blank titration)

Normality_Sulphuric - for Normality of standard sulphuric acid solution
}
\usage{
df_N_Boric
}
\description{
User is advised to prepare the data as suggested in the example to get the total N (when soil is digested with concentrated sulphuric acid in the presence of digestion mixture); or available or mineralizable N (when soil is distilled with alkaline potassium permanganate solution) in soil
}
\keyword{datasets}
