% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SoilTesting.R
\docType{data}
\name{df_N_Sulphuric}
\alias{df_N_Sulphuric}
\title{Example Data Frame for total nitrogen (N) in soil}
\format{
Data frame with Mass of soil sample (g) in the first column, Volume of standard sodium hydroxide solution (mL) consumed in titration with soil in the second column, Volume of standard sodium hydroxide solution (mL) consumed in titration without soil (i.e., blank titration) in the third column, and Normality of standard sodium hydroxide solution in the fourth column. Write the following notations on the spreadsheet:

Mass_Soil - for Mass of soil sample (g)

Vol_NaOH_Soil - for Volume of standard sodium hydroxide solution (mL) consumed in titration with soil

Vol_NaOH_Blank - for Volume of standard sodium hydroxide solution (mL) consumed in titration without soil (i.e., blank titration)

Normality_NaOH - for Normality of standard sodium hydroxide solution
}
\usage{
df_N_Sulphuric
}
\description{
User is advised to prepare the data as suggested in the example to get the total N in soil
}
\keyword{datasets}
