% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SoilTesting.R
\docType{data}
\name{df_Std_Curve}
\alias{df_Std_Curve}
\title{Example Data Frame for element concentration from standard curve}
\format{
Data frame with Concentration of the element in the working standard solutions (mg/L) in the first column, Absorbance or Emission reading of the working standard solutions in the second column, Absorbance or Emission reading of the sample solutions in the third column. Write the following notations on the spreadsheet:

Conc_mgL - for Concentration of the element in the working standard solutions (mg/L)

Abs_Emi_Std - for Absorbance or Emission reading of the working standard solutions

Abs_Emi_Sample - for Absorbance or Emission reading of the sample solutions
}
\usage{
df_Std_Curve
}
\description{
User is advised to prepare the data as suggested in the example to get the concentration of the element under consideration from absorbance or emission readings and elemental concentration in the working standard solutions.
}
\keyword{datasets}
