% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SoilTesting.R
\docType{data}
\name{df_WBC}
\alias{df_WBC}
\title{Example Data Frame for oxidizable organic carbon (C) or Walkley-Black C in soil}
\format{
Data frame with Mass of soil sample (g) in the first column, Volume of potassium dichromate solution (mL) in the second column, Volume of ferrous ammonium sulphate solution (mL) consumed in titration without soil (i.e., blank titration) in the third column, Volume of ferrous ammonium sulphate solution (mL) consumed in titration with soil in the fourth column, Normality of potassium dichromate solution in the fifth column. Write the following notations on the spreadsheet:

Mass_Soil - for Mass of soil sample (g)

Vol_Potassium_dichromate	-	for Volume of potassium dichromate solution (mL)

Vol_FAS_Blank - for Volume of ferrous ammonium sulphate solution (mL) consumed in titration without soil (i.e., blank titration)

Vol_FAS_Soil - for Volume of ferrous ammonium sulphate solution (mL) consumed in titration with soil

Normality_Potassium_dichromate - for Normality of potassium dichromate solution
}
\usage{
df_WBC
}
\description{
User is advised to prepare the data as suggested in the example to get the oxidizable organic C or Walkley-Black C (WBC) in soil
}
\keyword{datasets}
